requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
-e .
    # via -r requirements.in
aiofiles==22.1.0
    # via -r requirements.in
aioredis==2.0.1
    # via redis-om
altair==4.2.0
    # via streamlit
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
asttokens==2.1.0
    # via stack-data
async-timeout==4.0.2
    # via
    #   aioredis
    #   redis
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
blinker==1.5
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
cleo==1.0.0a5
    # via redis-om
click==8.1.3
    # via
    #   nltk
    #   redis-om
    #   streamlit
commonmark==0.9.1
    # via rich
contourpy==1.0.6
    # via matplotlib
crashtest==0.3.1
    # via cleo
cycler==0.11.0
    # via matplotlib
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   validators
deprecated==1.2.13
    # via redis
entrypoints==0.4
    # via
    #   altair
    #   jupyter-client
executing==1.2.0
    # via stack-data
filelock==3.8.0
    # via
    #   huggingface-hub
    #   transformers
flake8==3.8.3
    # via -r requirements.in
fonttools==4.38.0
    # via matplotlib
gitdb==4.0.9
    # via gitpython
gitpython==3.1.29
    # via streamlit
hiredis==2.0.0
    # via redis-om
huggingface-hub==0.10.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via streamlit
ipykernel==6.17.0
    # via
    #   ipywidgets
    #   pydeck
ipython==8.6.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.2
    # via pydeck
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
jsonschema==4.16.0
    # via altair
jupyter-client==7.4.4
    # via ipykernel
jupyter-core==4.11.2
    # via jupyter-client
jupyterlab-widgets==3.0.3
    # via ipywidgets
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.6.1
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mccabe==0.6.1
    # via flake8
more-itertools==8.14.0
    # via redis-om
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
networkx==2.8.7
    # via -r requirements.in
nltk==3.7
    # via sentence-transformers
numpy==1.23.4
    # via
    #   altair
    #   contourpy
    #   matplotlib
    #   pandas
    #   patsy
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   statsmodels
    #   streamlit
    #   torchvision
    #   transformers
packaging==21.3
    # via
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   redis
    #   statsmodels
    #   streamlit
    #   transformers
pandas==1.5.1
    # via
    #   -r requirements.in
    #   altair
    #   statsmodels
    #   streamlit
parso==0.8.3
    # via jedi
patsy==0.5.3
    # via statsmodels
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.3.0
    # via
    #   matplotlib
    #   streamlit
    #   torchvision
plotly==5.11.0
    # via -r requirements.in
pptree==3.1
    # via redis-om
prompt-toolkit==3.0.31
    # via ipython
protobuf==3.20.3
    # via streamlit
psutil==5.9.3
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==10.0.0
    # via streamlit
pycodestyle==2.6.0
    # via flake8
pydantic==1.10.2
    # via redis-om
pydeck==0.7.1
    # via streamlit
pyflakes==2.2.0
    # via flake8
pygments==2.13.0
    # via
    #   ipython
    #   rich
pylev==1.4.0
    # via cleo
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.19.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   streamlit
python-ulid==1.1.0
    # via redis-om
pytz==2022.5
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
redis==4.3.4
    # via redis-om
redis-om==0.1.0
    # via -r requirements.in
regex==2022.10.31
    # via
    #   nltk
    #   transformers
requests==2.28.1
    # via
    #   huggingface-hub
    #   streamlit
    #   torchvision
    #   transformers
rich==12.6.0
    # via streamlit
scikit-learn==1.1.3
    # via sentence-transformers
scipy==1.9.3
    # via
    #   -r requirements.in
    #   scikit-learn
    #   sentence-transformers
    #   statsmodels
semver==2.13.0
    # via streamlit
sentence-transformers==2.2.2
    # via -r requirements.in
sentencepiece==0.1.97
    # via sentence-transformers
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
smmap==5.0.0
    # via gitdb
stack-data==0.6.0
    # via ipython
statsmodels==0.13.5
    # via -r requirements.in
streamlit==1.14.0
    # via -r requirements.in
tenacity==8.1.0
    # via plotly
threadpoolctl==3.1.0
    # via scikit-learn
tokenizers==0.13.1
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
torch==1.13.0
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.14.0
    # via sentence-transformers
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   transformers
traitlets==5.5.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   pydeck
transformers==4.23.1
    # via sentence-transformers
types-redis==4.3.21.3
    # via redis-om
typing-extensions==4.4.0
    # via
    #   aioredis
    #   huggingface-hub
    #   pydantic
    #   redis-om
    #   streamlit
    #   torch
    #   torchvision
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.12
    # via requests
validators==0.20.0
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
widgetsnbextension==4.0.3
    # via ipywidgets
wrapt==1.14.1
    # via deprecated
zipp==3.10.0
    # via importlib-metadata
